luks.txt: add info about where to get gen_ekb.py etc.
It is not obvious from the documentation where to find the sources for
the `luks-srv` and `hwkey-agent` trusty apps nor the `gen_ekb.py`
script. This file is just some bread crumbs to help find them.

This is not about secureboot, but is related to security and users might
find it useful.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>